New GC, more benchmarks (#482)
* try out earley bench

* checkpoint

* checkpoint

* update dep

* gc notes

* free list stuff

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* go back to older implementation

* more checkpoint

* cursor is back working but needs some love

* more debugging

* checkpoint yet again

* checkpoint

* finally things don't explode

* checkpoint again

* checkpoint

* more perf stuff

* checkpoint

* checkpoint

* weak boxes

* weak box stuff

* checkpoint again

* checkpoint

* lattice benchmark

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* checkpoint

* more benchmarks

* checkpoint

* checkpoint

* checkpoint

* checkpoint on silly reader hacks to get things working better

* another benchmark

* checkpoint

* include profiler

* checkpoint

* reduce allocation in marking

* checkpoint

* checkpoint

* add op codes for statically verifiable checks when relevant

* some more comments

* checkpoint

* checkpoint

* shrink initial heap size

* clean up

* get working on non sync

* update lists lib

* clean up

* fix bug with map

* set default for mimalloc for now

* fix eval bug

* raise actions timeout

* go back to mimalloc

* increase timeout

* add self tail call back

* fix the gc, more tests

* more tests

* more tests

* more tests

+++ crates/steel-parser/src/tokens.rs
@@ -457,7 +457,7 @@ impl<'a, TY> TokenLike<'a, TY> {
     pub const fn new(
         ty: TY,
         source: &'a str,
-        range: ops::Range<usize>,
+        range: ops::Range<u32>,
         source_id: Option<SourceId>,
     ) -> Self {
         Self {
@@ -479,7 +479,7 @@ impl<'a, T> Token<'a, T> {
         self.span
     }
 
-    pub const fn range(&self) -> ops::Range<usize> {
+    pub const fn range(&self) -> ops::Range<u32> {
         self.span.start()..self.span.end()
     }
 
@@ -500,37 +500,37 @@ impl<T> From<&Token<'_, T>> for Span {
     }
 }
 
-impl<T> From<Token<'_, T>> for ops::Range<usize> {
+impl<T> From<Token<'_, T>> for ops::Range<u32> {
     fn from(token: Token<'_, T>) -> Self {
         token.span().into()
     }
 }
 
-impl<T> From<&Token<'_, T>> for ops::Range<usize> {
+impl<T> From<&Token<'_, T>> for ops::Range<u32> {
     fn from(token: &Token<'_, T>) -> Self {
         token.span().into()
     }
 }
 
-impl<T> From<Token<'_, T>> for (usize, usize) {
+impl<T> From<Token<'_, T>> for (u32, u32) {
     fn from(token: Token<'_, T>) -> Self {
         token.span().into()
     }
 }
 
-impl<T> From<&Token<'_, T>> for (usize, usize) {
+impl<T> From<&Token<'_, T>> for (u32, u32) {
     fn from(token: &Token<'_, T>) -> Self {
         token.span().into()
     }
 }
 
-impl<T> From<Token<'_, T>> for [usize; 2] {
+impl<T> From<Token<'_, T>> for [u32; 2] {
     fn from(token: Token<'_, T>) -> Self {
         token.span().into()
     }
 }
 
-impl<T> From<&Token<'_, T>> for [usize; 2] {
+impl<T> From<&Token<'_, T>> for [u32; 2] {
     fn from(token: &Token<'_, T>) -> Self {
         token.span().into()
     }
